remove old gcc-10 env vars, since the newer ubuntu machine doesn't ship with that.
also, these env vars were never a real solution to pinning the glibc version, it was just lucky for a few saleae analyzers.

for now, CI builds might not run on older linux distros, due to the possibility that a newer glibc version could be needed.

The only real solution to this will be to use a container with an older set of libraries.